move userExistence store and some user-related functions into separate directory
