+++ jsapp/js/assetUtils.ts
@@ -6,7 +6,7 @@
 import React from 'react';
 import {stores} from 'js/stores';
 import permConfig from 'js/components/permissions/permConfig';
-import {buildUserUrl} from 'js/utils';
+import {ANON_USERNAME_URL} from 'js/users/utils';
 import envStore from 'js/envStore';
 import sessionStore from 'js/stores/session';
 import type {
@@ -24,7 +24,6 @@ import {
   GROUP_TYPES_END,
   SCORE_ROW_TYPE,
   RANK_LEVEL_TYPE,
-  ANON_USERNAME,
   PERMISSIONS_CODENAMES,
   ACCESS_TYPES,
   ROOT_URL,
@@ -256,7 +255,7 @@ export function isAssetPublic(permissions?: PermissionResponse[]) {
   permissions?.forEach((perm) => {
     const foundPerm = permConfig.getPermissionByCodename(PERMISSIONS_CODENAMES.discover_asset);
     if (
-      perm.user === buildUserUrl(ANON_USERNAME) &&
+      perm.user === ANON_USERNAME_URL &&
       foundPerm !== undefined &&
       perm.permission === foundPerm.url
     ) {
